analyzer/requirements.txt
flask==2.3.3
androguard==3.3.5
werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.9.0
python-docx==0.8.11
